+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+## 🐜 62.3.0 - 10-11-2023
+
 ### Added
 * Document Component: Mark en highlight functionaliteit ([#2386](https://github.com/dso-toolkit/dso-toolkit/issues/2386))
 * Date Picker: vorige implementatie herintroduceren als "Date Picker Legacy" ([#2430](https://github.com/dso-toolkit/dso-toolkit/issues/2430))